requirements.txt
uptime-kuma-api
Flask
python-dotenv
flask[async]